"Refactor user model import path in multiple files"
This commit refactors the import path of the User model in multiple files from 'src/users/users.model' to 'src/users/model/users.model'.
