+++ src/auth/auth.service.ts
@@ -8,7 +8,7 @@ import { JwtService } from '@nestjs/jwt'
 import { CreateUserDto } from 'src/users/dto/create-user.dto'
 import { UsersService } from 'src/users/users.service'
 import * as bcrypt from 'bcryptjs'
-import { User } from 'src/users/users.model'
+import { User } from 'src/users/model/users.model'
 @Injectable()
 export class AuthService {
     constructor(
