+++ src/posts/posts.model.ts
@@ -10,7 +10,7 @@ import {
     Model,
     Table,
 } from 'sequelize-typescript'
-import { User } from 'src/users/users.model'
+import { User } from 'src/users/model/users.model'
 
 interface PostCreationAttrs {
     title: string
