+++ src/roles/roles.module.ts
@@ -3,7 +3,7 @@ import { RolesService } from './roles.service'
 import { RolesController } from './roles.controller'
 import { SequelizeModule } from '@nestjs/sequelize'
 import { Role } from './roles.model'
-import { User } from '../users/users.model'
+import { User } from '../users/model/users.model'
 import { UserRoles } from './user-roles.model'
 
 @Module({
