+++ src/roles/user-roles.model.ts
@@ -7,7 +7,7 @@ import {
     Model,
     Table,
 } from 'sequelize-typescript'
-import { User } from 'src/users/users.model'
+import { User } from 'src/users/model/users.model'
 import { Role } from './roles.model'
 
 @Table({ tableName: 'user-roles', timestamps: false })
