+++ src/users/model/users.model.ts
