+++ src/users/users.controller.ts
@@ -10,7 +10,7 @@ import {
 import { CreateUserDto } from './dto/create-user.dto'
 import { UsersService } from './users.service'
 import { ApiOperation, ApiResponse, ApiTags } from '@nestjs/swagger'
-import { User } from './users.model'
+import { User } from './model/users.model'
 import { JwtAuthGuard } from 'src/auth/jwt-auth.guard'
 import { RolesGuard } from 'src/auth/roles.guard'
 import { Roles } from 'src/auth/roles-auth.decorator'
