+++ src/users/users.module.ts
@@ -1,7 +1,7 @@
 import { Module, forwardRef } from '@nestjs/common'
 import { UsersController } from './users.controller'
 import { UsersService } from './users.service'
-import { User } from './users.model'
+import { User } from './model/users.model'
 import { SequelizeModule } from '@nestjs/sequelize'
 import { Role } from 'src/roles/roles.model'
 import { RolesModule } from 'src/roles/roles.module'
