+++ src/users/users.service.ts
@@ -1,5 +1,5 @@
 import { HttpStatus, Injectable, HttpException } from '@nestjs/common'
-import { User } from './users.model'
+import { User } from './model/users.model'
 import { InjectModel } from '@nestjs/sequelize'
 import { CreateUserDto } from './dto/create-user.dto'
 import { RolesService } from '../roles/roles.service'
